LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/path-objects/2d.path.bezierCurveTo.nonfinite.worker-expected.txt


PASS bezierCurveTo() with Infinity/NaN is ignored
